sync init
