+++ backend/api/quivr_api/models/settings.py
@@ -1,13 +1,9 @@
-from typing import Optional
 from uuid import UUID
 
 from posthog import Posthog
 from pydantic_settings import BaseSettings, SettingsConfigDict
-from sqlalchemy import Engine
 
 from quivr_api.logger import get_logger
-from quivr_api.models.databases.supabase.supabase import SupabaseDB
-from supabase.client import AsyncClient, Client
 
 logger = get_logger(__name__)
 
@@ -135,12 +131,5 @@ class ResendSettings(BaseSettings):
     quivr_smtp_username: str = ""
     quivr_smtp_password: str = ""
 
-
-# Global variables to store the Supabase client and database instances
-_supabase_client: Optional[Client] = None
-_supabase_async_client: Optional[AsyncClient] = None
-_supabase_db: Optional[SupabaseDB] = None
-_db_engine: Optional[Engine] = None
-_embedding_service = None
 
 settings = BrainSettings()  # type: ignore
